WIP adding xterm to playground
