+++ packages/wasi-motor/programs/src/bin/echo.rs
@@ -1,7 +1,7 @@
 use std::env;
 
 fn main() {
-    for env in env::args().skip(1) {
-        print!("{} ", env);
+    for arg in env::args().skip(1) {
+        print!("{} ", arg);
     }
 }
