+++ packages/website/package.json
@@ -21,18 +21,20 @@
     "serve": "vite preview"
   },
   "devDependencies": {
+    "@runno/host": "^0.3.0",
+    "@runno/runtime": "^0.3.1",
+    "@runno/wasi-motor": "^0.3.0",
     "@tailwindcss/typography": "^0.5.4",
     "autoprefixer": "^10.4.0",
     "lit": "^2.3.1",
     "postcss": "^8.4.16",
     "tailwindcss": "^3.1.8",
     "typescript": "^4.8.2",
-    "vite": "^2.6.13"
+    "vite": "^2.6.13",
+    "xterm": "^4.19.0",
+    "xterm-addon-fit": "^0.5.0",
+    "xterm-addon-web-links": "^0.6.0"
   },
-  "dependencies": {
-    "@runno/host": "^0.3.0",
-    "@runno/runtime": "^0.3.1",
-    "@runno/wasi-motor": "^0.3.0"
-  },
+  "dependencies": {},
   "gitHead": "812d21ec7c0c5d57c0274caaf7ff21c79ca79c1f"
 }
